Consistency - pcts and angles always floats

+++ sources/class-CCSeries.php
@@ -35,13 +35,13 @@ class CCSeries extends EventSeries
 	 * @param int $controller - The MIDI controller #
 	 * @param int $shape - The shape of the curve
 	 * @param float $freq - The frequency, # of cycles per duration
-	 * @param int $offset - The angle of offset, passed in degrees
-	 * @param int $min_pct - Minimum value used in scaling, in percent
-	 * @param int $max_pct - Maximum value used in scaling, in percent
+	 * @param float $offset - The angle of offset, passed in degrees
+	 * @param float $min_pct - Minimum value used in scaling, in percent
+	 * @param float $max_pct - Maximum value used in scaling, in percent
 	 * @param int $tick_inc - How far apart in ticks to spread the events
 	 * @return void
 	 */
-	function __construct(int $chan = 0, int $controller = 11, int $shape = EVENTSeries::SINE, float $freq = 1, int $offset = 0, int $min_pct = 0, int $max_pct = 100, int $tick_inc = 48)
+	function __construct(int $chan = 0, int $controller = 11, int $shape = EVENTSeries::SINE, float $freq = 1, float $offset = 0, float $min_pct = 0, float $max_pct = 100, int $tick_inc = 48)
 	{
 		$this->channel = MIDIEvent::rangeCheck($chan, 0, 0xF);
 		$this->controller = MIDIEvent::rangeCheck($controller);
